refactor: refactore some components
