+++ tels-crm-project/src/component/truckCard/TruckCard.tsx
@@ -1,19 +1,18 @@
 import { LinearProgress } from "@mui/material";
 import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
-import Flag from "react-world-flags";
-import TruckType from "../../types/truckType";
 import { useActions } from "../hooks/useActions";
 import { useSelector } from "../hooks/useSelector";
 import useTranslate from "../hooks/useTranslate";
 import Map from "./map/Map";
+import TruckInfo from "./truckInfo/TruckInfo";
 
 import c from "./TruckCard.module.scss";
 
 const TruckCard: React.FC = () => {
     const { id } = useParams();
 
-    const { lang, t } = useTranslate();
+    const { t } = useTranslate();
 
     const data = useSelector((state) => state.truck.data);
     const loading = useSelector((state) => state.truck.loading);
@@ -46,34 +45,7 @@ const TruckCard: React.FC = () => {
                 {t("truckCard.tittle")} {data[0].auto_number}
             </h2>
             <div className={c.truckInfo}>
-                <ul>
-                    <li className={c.number}>
-                        <span>{t("trucks.number")}: </span>
-                        {data[0].auto_number}
-                    </li>
-                    <li className={c.country}>
-                        <span>{t("truckCard.country")}:</span>
-                        <Flag code={data[0].country_shot} height="16" />
-                    </li>
-                    <li className={c.phone}>
-                        <span>{t("trucks.phone")}: </span>+
-                        {data[0].phone_number}
-                    </li>
-                    <li className={c.speed}>
-                        <span>{t("trucks.speed")}: </span>
-                        {data[0].speed_can}
-                    </li>
-                    <li className={c.date}>
-                        <span>{t("trucks.date")}: </span>
-                        {data[0].datetime}
-                    </li>
-                    <li>
-                        <span>{t("truckCard.adress")}</span>
-                        <div className={c.place}>
-                            {lang === "ru" ? data[0].place : data[0].place_en}
-                        </div>
-                    </li>
-                </ul>
+                <TruckInfo data={data[0]} />
                 <Map
                     latitude={data[0].latitude}
                     longitude={data[0].longitude}
