+++ tels-crm-project/src/component/truckCard/map/Map.tsx
@@ -1,61 +1,3 @@
-// import React, { memo, SetStateAction, useCallback, useState } from "react";
-// import { GoogleMap, Marker, useJsApiLoader } from "@react-google-maps/api";
-
-// const containerStyle = {
-//     width: "500px",
-//     height: "500px",
-//     margin: "40px",
-// };
-
-// type PropsType = {
-//     latitude?: string;
-//     longitude?: string;
-// };
-
-// const Map: React.FC<PropsType> = ({
-//     latitude = "55.018600",
-//     longitude = "32.293196",
-// }) => {
-//     const center = {
-//         lat: +latitude,
-//         lng: +longitude,
-//     };
-
-//     const { isLoaded } = useJsApiLoader({
-//         id: "google-map-script",
-//         googleMapsApiKey: "AIzaSyCXDhAfeFG_kqxjh_l219GRV2VHGzM2j4w",
-//     });
-
-//     const [map, setMap] = useState(null);
-
-//     const onLoad = useCallback(function callback(map: google.maps.Map) {
-//         const bounds = new window.google.maps.LatLngBounds(center);
-//         map.fitBounds(bounds);
-//         // setMap(map);
-//     }, []);
-
-//     const onUnmount = useCallback(function callback(map: google.maps.Map) {
-//         setMap(null);
-//     }, []);
-
-//     return isLoaded ? (
-//         <GoogleMap
-//             mapContainerStyle={containerStyle}
-//             center={center}
-//             zoom={1}
-//             onLoad={onLoad}
-//             onUnmount={onUnmount}
-//         >
-//             <Marker />
-//             <></>
-//         </GoogleMap>
-//     ) : (
-//         <></>
-//     );
-// };
-
-// export default memo(Map);
-
 import React, { useState } from "react";
 import GoogleMapReact from "google-map-react";
 import Marker from "./Marker";
@@ -71,7 +13,7 @@ const Map: React.FC<PropsType> = ({
     longitude = "32.293196",
     number = "number",
 }) => {
-    const getMapOptions = (maps: any) => {
+    const getMapOptions = () => {
         return {
             disableDefaultUI: true,
             mapTypeControl: true,
