+++ tels-crm-project/src/component/ui/button/Button.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import c from "./Button.module.scss";
 
 type PropsType = {
-    onClick: (e: React.MouseEvent<HTMLButtonElement>) => void;
+    onClick?: (e: React.MouseEvent<HTMLButtonElement>) => void;
     children: React.ReactNode;
     pulse?: boolean;
 };
